Rename schedule controls to timeout
