+++ assets/data-port/export/store/index.js
@@ -1,6 +1,6 @@
 import { registerStore } from '@wordpress/data';
 import { controls as dataControls } from '@wordpress/data-controls';
-import scheduleControls from '../../../shared/data/schedule-controls';
+import scheduleControls from '../../../shared/data/timeout-controls';
 import { createReducerFromActionMap } from '../../../shared/data/store-helpers';
 
 import * as actions from './actions';
